+++ src/texture_data.rs
@@ -40,11 +40,11 @@ impl<'a> TextureData{
             }
         }
     }
-    pub fn draw(&self, canvas : &mut sdl2::render::Canvas<sdl2::video::Window>, texture_map : &std::collections::HashMap<String,Texture>, x : i32, y : i32, width : u32, height : u32) -> Result<(),String>{
+    pub fn draw(&self, canvas : &mut sdl2::render::Canvas<sdl2::video::Window>, texture_map : &std::collections::HashMap<String,Texture>, x : f64, y : f64, width : u32, height : u32) -> Result<(),String>{
         match texture_map.get(&self.path.clone()){
             Some(texture) => {
                // println!("self x: {}, self y: {}, self width: {}, self height: {}",self.x,self.y,self.width,self.height);
-                canvas.copy(texture,sdl2::rect::Rect::new(self.x,self.y,self.width,self.height),sdl2::rect::Rect::new(x,y,width,height)).map_err(|e| e.to_string())
+                canvas.copy(texture,sdl2::rect::Rect::new(self.x,self.y,self.width,self.height),sdl2::rect::Rect::new(x as i32,y as i32,width,height)).map_err(|e| e.to_string())
             },
             None => Err("Texture not loaded".into())
         }
